blockbuilder/scheduler: move partitionState into its own file
Move the partitionState and jobState types, bucket constants, and the
partitionState methods out of scheduler.go into partition_state.go, and
the TestPartitionState* tests into partition_state_test.go.
